Refactor: complete modularization, loot pipeline, and cleanup
- Modularize into domain services: Common (State, ConfigLoader, PathResolver), Containers (EmptyBoosterFactory, SecureContainerFilterUpdater, PouchCompatibilityUpdater), Cards (CardPriceResolver, CardItemFactory), Binders (BinderFactory), Traders (TraderOfferService, FenceService), Ragfair (RagfairConfigurator), Loot (LootService, LootInjector, ConstantsContainer)

- Move constants to Services/Loot; remove duplicate

- Simplify PostDb to orchestrate modules; update DI wiring

- Reduce log noise; remove 100% spawn debug path; rename Try* to success-based names

- Update imports/namespaces; fix ConfigServer references

- Build passes; 1 existing nullability warning in TTCModMetadata
